+++ app/src/androidTest/java/com/emenjivar/luminar/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.emenjivar.camerafilter
+package com.emenjivar.luminar
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.emenjivar.camerafilter", appContext.packageName)
+        assertEquals("com.emenjivar.luminar", appContext.packageName)
     }
 }
