+++ app/src/main/java/com/emenjivar/luminar/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.emenjivar.camerafilter
+package com.emenjivar.luminar
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -7,8 +7,8 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import com.emenjivar.camerafilter.screen.camera.CameraScreen
-import com.emenjivar.camerafilter.ui.theme.RealTimeCameraFilterTheme
+import com.emenjivar.luminar.screen.camera.CameraScreen
+import com.emenjivar.luminar.ui.theme.RealTimeCameraFilterTheme
 import org.opencv.android.OpenCVLoader
 
 class MainActivity : ComponentActivity() {
