+++ app/src/main/java/com/emenjivar/luminar/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.emenjivar.camerafilter.ui.theme
+package com.emenjivar.luminar.ui.theme
 
 import androidx.compose.ui.graphics.Color
