+++ app/src/main/java/com/emenjivar/luminar/ui/widget/FilterBubble.kt
@@ -1,4 +1,4 @@
-package com.emenjivar.camerafilter.ui.widget
+package com.emenjivar.luminar.ui.widget
 
 import android.graphics.Bitmap
 import android.widget.ImageView
