+++ app/src/test/java/com/emenjivar/luminar/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.emenjivar.camerafilter
+package com.emenjivar.luminar
 
 import org.junit.Test
